refactor(users): replace useUsersAdmin with individual hooks for better modularity and clarity
