+++ src/modules/admin/users/hooks/useRegisterUser.ts
@@ -1,0 +1,20 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useApiRequest } from '@/hooks/useApiRequest';
+
+export const useRegisterUser = () => {
+  const { apiRequest } = useApiRequest();
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: async (data: any) =>
+      apiRequest({
+        url: '/admin/users/register',
+        method: 'post',
+        data,
+        requireAuth: true,
+      }),
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['users'] });
+    },
+  });
+};
